+++ lib/pleroma/api/notification/pleroma_api_notification_service_impl.dart
@@ -104,7 +104,9 @@ class PleromaApiNotificationService extends BasePleromaApiService
     List<PleromaApiNotificationType>? includeTypes,
     List<PleromaApiVisibility>? excludeVisibilities,
   }) async {
-    _checkGetNotificationsIncludeTypes(includeTypes);
+    if(restApiAuthService.isPleroma) {
+      _checkGetNotificationsIncludeTypes(includeTypes);
+    }
 
     _checkGetNotificationsExcludeVisibilities(excludeVisibilities);
     excludeTypes = _checkGetNotificationsExcludeTypes(excludeTypes);